+++ ArcDPS Boon Table/Boon.h
@@ -7,7 +7,7 @@
 class Boon
 {
 public:
-	int32_t duration = 0;
+	uint64_t duration = 0;
 	BoonDef* def = nullptr;
 	uint64_t expected_end_time = 0;
 	
@@ -18,8 +18,8 @@ public:
 	void Apply(int32_t new_duration);
 	void Remove(int32_t new_duration);
 
-	int32_t getDuration(uint64_t new_current_time);
+	uint64_t getDuration(uint64_t new_current_time);
 	uint64_t getDurationRemaining(uint64_t new_current_time);
 
-	float getUptime(uint64_t new_current_time, uint64_t new_combat_time);
+	double getUptime(uint64_t new_current_time, uint64_t new_combat_time);
 };
